requirements.txt
python-jose
passlib
uvicorn[standard]
SQLAlchemy
MySQL-python
pydantic
fastapi